Add base and WEP power displaying, fix suspended armament adding
Suspended armament was throwing an integrity error, because first query without stripped postfix always returned none.
When the script stripped the bomb/rocket etc. postfix in the given arm_name it could later
appear that the object is already in the database, but couldn't be found because the name in the
database had stripped postfix.